images/memorystore-redis/data-gen/requirements.txt
Faker
google-auth
google-cloud-redis
redis
requests

